+++ testsuite/td-gauss-sldae-a1/predefines.h
@@ -1,13 +1,13 @@
 /**
  * Define lattice size and lattice spacing.
  * */
-#define NX 8
-#define NY 10
+#define NX 12
+#define NY 12
 #define NZ 12
 
-#define DX 1.0
-#define DY 1.0
-#define DZ 1.0
+#define DX 1.15
+#define DY 1.15
+#define DZ 1.15
 
 /**
  * Select functional:
@@ -19,31 +19,21 @@
  *      for simulating unitary Fermi gas,
  *      at qualitative level it produces results compatible with SLDA, however it is more accurate,
  *      due to presence of current terms in the functional it has worse convergence properties.
- *  - SLDAE:
- *      for simulating Fermi gas for an arbitrary value of akF,
- *      for small and negative akF the functional is equivalent to BDG, while for large akF is equivalent to ASLDA.
- *      To speed-up computation, you can use SLDAE_FORCE_A1 option.
- *      It sets effective mass=1, which renders the current dependence of the functional. 
- *      For more info see: https://arxiv.org/abs/2201.07626
  *  - BDG:
  *      for simulating systems in BCS regime,
  *      equations of motion are equivalent to Bogoliubov-de-Gennes equations,
- *  - CUSTOMEDF:
- *      use this option to define your custom functional,
- *      then you need to provide body of functions: tdwslda_compute_energy( ) and tdwslda_compute_potentials( )
- *      in problem-definition.h file
+ *      you MUST set aBdG value in input file when using this functional.
  * */
-#define FUNCTIONAL SLDA
+// #define FUNCTIONAL SLDA
 // #define FUNCTIONAL ASLDA
-// #define FUNCTIONAL SLDAE
 // #define FUNCTIONAL BDG
-// #define FUNCTIONAL CUSTOMEDF
+#define FUNCTIONAL SLDAE
 
 /**
  * Meaningful only in case SLDAE.
  * Sets effective mass to be equal, and speeds-up computation (approximately by a factor of two)
  * */
-// #define SLDAE_FORCE_A1
+#define SLDAE_FORCE_A1
 
 /**
  * Select which external potentials you want to use in simulations.
@@ -53,14 +43,11 @@
  *       function delta_ext(...) from problem definition will be called in each iteration.
  *   - ENABLE_VELOCITY_EXT:
  *       function velocity_ext(...) from problem definition will be called in each iteration.
- *   - ENABLE_MODIFY_POTENTIALS:
- *       function modify_potentials(...) from problem definition will be called in each iteration.
  * In order to achieve best performance disable call of empty functions.
  * */
 #define ENABLE_V_EXT
 // #define ENABLE_DELTA_EXT
 // #define ENABLE_VELOCITY_EXT
-// #define ENABLE_MODIFY_POTENTIALS
 
 /**
  * Maximal number of parameters in params array
@@ -78,7 +65,7 @@
  * In such case, the evolves only wave-functions for single spin component
  * and in consequence computing time decreases by factor of two
  * */
-// #define SPINSYMMETRY_MODE
+#define SPINSYMMETRY_MODE
 
 /**
  * Scheme of pairing field renormalization procedure. 
@@ -87,11 +74,11 @@
  * SPHERICAL_CUTOFF: use spherical momentum space cutoff, in this case you need to set `ec` variable in input file (default).
  * CUBIC_CUTOFF: use cubic momentum space cutoff, in this case `ec` will be set to infinity automatically.
  * */
-#define REGULARIZATION_SCHEME SPHERICAL_CUTOFF
+// #define REGULARIZATION_SCHEME SPHERICAL_CUTOFF
 // #define REGULARIZATION_SCHEME CUBIC_CUTOFF
 
 /**
- * Meaningful only in case of ASLDA and SLDAE.
+ * Meaningful only in case of ASLDA.
  * Parameters defining stabilization procedure of ASLDA functional. 
  * For regions with density smaller than ASLDA_STABILIZATION_EXCLUDE_BELOW_DENISTY 
  * contribution from current term j^2/2n is assumed to be zero. 
@@ -104,18 +91,69 @@
 #define ASLDA_STABILIZATION_EXCLUDE_BELOW_DENISTY 1.0e-7
 
 /**
+ * Number of mpi processes per IO group used for collective (parallel) writing of checkpoint files.
+ * Performance of read/write checkpoint depends on the number of writes involved in IO process,
+ * and optimal value depends on the computer. 
+ * Use the default value (24) unless you are not satisfied with IO performance. 
+ * */
+#define MPI_NP_PER_IO_GROUP 24
+
+/**
  * Active this flag in order to store quasi-particle energies for each measurement.
  * Note that in case of 1d or 2d codes it can require much more space than measurements itself.
  * */
 // #define STORE_QPE
 
 /**
- * Machine file. 
- * This file contains info about machine that will be used in the computation process.
- * You can specify the file in the following ways: 
- * - copy `machine.h` file to the current directory, see templates folder for various examples,
- * - specify the folder with `machine.h` file via -I option in Makefile  
- * - use system variable WSLDA_MACHINE to specify the folder with `machine.h` file, for example:
- *   export WSLDA_MACHINE=...
- **/ 
-#include "machine.h"
+ * Activate this flag in order to print to stdout
+ * applied mapping mpi-process <==> device-id.
+ * */
+#define PRINT_GPU_DISTRIBUTION
+
+/**
+ * Activate this flag if target machine has non-standard distribution of GPUs. 
+ * In such case you need to provide body of function `assign_deviceid_to_mpi_process`.
+ * If this flag is commented-out it is assumed that code is running on a machine 
+ * with uniformly distributed GPU cards across the nodes, 
+ * and each node has `gpuspernode` (input file parameter) cards.
+ * */
+#define CUSTOM_GPU_DISTRIBUTION
+
+/**
+ * This function is used to assign unique device-id to mpi process.
+ * @param comm MPI communicator
+ * @return device-id assign to the process extracted by function MPI_Comm_rank(...)
+ * DO NOT REMOVE STATEMENT `#if ... BELOW !!!
+ * */
+#if defined(CUSTOM_GPU_DISTRIBUTION) && defined(TDWSLDA_MAIN)
+int assign_deviceid_to_mpi_process(MPI_Comm comm)
+{
+    int np, ip;
+    MPI_Comm_size(comm, &np);
+    MPI_Comm_rank(comm, &ip);
+    
+    // assign here deviceid to process with ip=iam
+    int deviceid=0;
+    
+    if(ip==0) printf("# CUSTOM GPU DISTRIBUTION FOR MACHINE: DWARF\n");
+    char processor_name[MPI_MAX_PROCESSOR_NAME];
+    int name_len;
+    MPI_Get_processor_name(processor_name, &name_len);
+
+    int ompi_ppn=4;
+    if(strcmp (processor_name,"node2061.grid4cern.if.pw.edu.pl")==0) ompi_ppn=8;
+    if(strcmp (processor_name,"node2062.grid4cern.if.pw.edu.pl")==0) ompi_ppn=8;
+    if(strcmp (processor_name,"node2067.grid4cern.if.pw.edu.pl")==0) ompi_ppn=8;
+    if(strcmp (processor_name,"node2068.grid4cern.if.pw.edu.pl")==0) ompi_ppn=2;
+
+
+    deviceid=ip % 8;
+    
+    return deviceid % ompi_ppn;
+}
+#endif
+
+// setting code in testing mode with uniform system
+#define UNIFORM_TEST_MODE
+
+