+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 2022.xx.xx
-* Update of SLDAE functional (redefinition of the effective mass, now A=alpha)
+* Update of SLDAE functional (redefinition of the effective mass, now A=alpha; added a new flag that forces the effective mass to be 1.0)
 
 # 2022.01.05
 * A new functional SLDAE has been implemented.